fix proposal freelancer name

+++ backend/project/views.py
@@ -28,32 +28,54 @@ class ProjectListCreateView(APIView):
   
 
 class ProjectDetailView(APIView):
-  """
-  API endpoint for retrieving, updating, and deleting projects
-  """
+  
   def get_object(self, pk):
     try:
       return Project.objects.get(pk=pk)
     except Project.DoesNotExist:
       return Response(status=status.HTTP_404_NOT_FOUND)
 
+  # def get(self, request, pk):
+  #   project = self.get_object(pk)
+  #   serializer = ProjectSerializer(project)
+  #   return Response(serializer.data)
+  
   def get(self, request, pk):
-    project = self.get_object(pk)
-    serializer = ProjectSerializer(project)
-    return Response(serializer.data)
+        project = self.get_object(pk)
+        if not isinstance(project, Project):
+            return project  # If project is a Response, it means 404 error
+        serializer = ProjectSerializer(project)
+        return Response(serializer.data)
 
+  # def put(self, request, pk):
+  #   project = self.get_object(pk)
+  #   serializer = ProjectSerializer(project, data=request.data)
+  #   if serializer.is_valid():
+  #     serializer.save()
+  #     return Response(serializer.data)
+  #   return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+  
   def put(self, request, pk):
-    project = self.get_object(pk)
-    serializer = ProjectSerializer(project, data=request.data)
-    if serializer.is_valid():
-      serializer.save()
-      return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        project = self.get_object(pk)
+        if not isinstance(project, Project):
+            return project  
+        serializer = ProjectSerializer(project, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+  # def delete(self, request, pk):
+  #   project = self.get_object(pk)
+  #   project.delete()
+  #   return Response(status=status.HTTP_204_NO_CONTENT)
+  
   def delete(self, request, pk):
-    project = self.get_object(pk)
-    project.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+        project = self.get_object(pk)
+        if not isinstance(project, Project):
+            return project  
+        project.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
